# Speed up isbitsunion and bitsunionsize functions