@quinnj 2017-09-06
@vtjnash, any concerns w/ this approach?