@vtjnash 2017-09-06
This is OK, but doesn't implement the same function anymore. In particular, while the `jl_islayout_inline` query is related to `isbitsunion`, it is not correct to assume the queries are equivalent.